Code changes for T2 telemetry markers
