+++ source/TR-181/middle_layer_src/wanmgr_dml_iface_v2_apis.c
@@ -2070,6 +2070,15 @@ BOOL WanIfIpCfg_SetParamUlongValue(ANSC_HANDLE hInsContext, char* ParamName, ULO
                 p_VirtIf->IP.RefreshDHCP = TRUE;
                 p_VirtIf->IP.ModeForceEnable = TRUE;
                 CcspTraceInfo(("%s %d IP.%s changed for %s to %d. ModeForceEnable set and Refreshing DHCP \n", __FUNCTION__, __LINE__, ParamName, p_VirtIf->Name, p_VirtIf->IP.Mode));
+                //Telemetry start
+                WanMgr_Telemetry_Marker_t Marker = {0};
+                Marker.enTelemetryMarkerID = WAN_INFO_IP_MODE;
+                Marker.pVirtInterface = p_VirtIf ;
+                if(ANSC_STATUS_FAILURE == wanmgr_telemetry_event(&Marker)){
+                    CcspTraceError(("%s %d: Error sending Telemetry event WAN_INFO_IP_MODE..\n",__FUNCTION__, __LINE__));
+                }
+                CcspTraceInfo(("%s %d: KAVYA, WAN_INFO_IP_MODE..\n",__FUNCTION__, __LINE__));
+                //Telemetry end		
             }
             ret = TRUE;
         }
